--- dojo/db_migrations/0023_cve_fix_1553.py
@@ -1,24 +1,0 @@
-# Generated by Django 2.2.4 on 2019-12-12 18:17
-
-import django.core.validators
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('dojo', '0023_SAST_track_unique_vulnerabilities'),
-    ]
-
-    operations = [
-        migrations.AlterField(
-            model_name='finding',
-            name='cve',
-            field=models.TextField(max_length=28, null=True, validators=[django.core.validators.RegexValidator(message="Vulnerability ID must be entered in the format: 'ABC-9999-9999'. ", regex='^[A-Z]{1,10}-\\d{4}-\\d{4,12}$')]),
-        ),
-        migrations.AlterField(
-            model_name='finding_template',
-            name='cve',
-            field=models.TextField(max_length=28, null=True, validators=[django.core.validators.RegexValidator(message="Vulnerability ID must be entered in the format: 'ABC-9999-9999'. ", regex='^[A-Z]{1,10}-\\d{4}-\\d{4,12}$')]),
-        ),
-    ]
