+++ dojo/db_migrations/0024_cve_regex_fix.py
@@ -1,0 +1,25 @@
+# Generated by Django 2.2.4 on 2019-12-13 13:54
+
+import django.core.validators
+from django.db import migrations, models
+import re
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('dojo', '0023_cve_fix_1553'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='finding',
+            name='cve',
+            field=models.TextField(max_length=28, null=True, validators=[django.core.validators.RegexValidator(re.compile('^[A-Z]{1,10}-\\d{4}-\\d{4,12}$'), message='Vulnerability ID must comply with regex ^[A-Z]{1,10}-\\d{4}-\\d{4,12}$')]),
+        ),
+        migrations.AlterField(
+            model_name='finding_template',
+            name='cve',
+            field=models.CharField(max_length=28, null=True, validators=[django.core.validators.RegexValidator(re.compile('^[A-Z]{1,10}-\\d{4}-\\d{4,12}$'), message='Vulnerability ID must comply with regex ^[A-Z]{1,10}-\\d{4}-\\d{4,12}$')]),
+        ),
+    ]
